+++ src/main/kotlin/ru/mipt/npm/gradle/KScienceCommonPlugin.kt
@@ -20,11 +20,6 @@ open class KScienceCommonPlugin : Plugin<Project> {
         registerKScienceExtension()
         repositories.applyRepos()
 
-        // apply dokka for all projects
-        if (!plugins.hasPlugin("org.jetbrains.dokka")) {
-            plugins.apply("org.jetbrains.dokka")
-        }
-
         //Configuration for K-JVM plugin
         pluginManager.withPlugin("org.jetbrains.kotlin.jvm") {
             //logger.info("Applying KScience configuration for JVM project")
@@ -43,6 +38,22 @@ open class KScienceCommonPlugin : Plugin<Project> {
                     }
                 }
             }
+            tasks.withType<KotlinJvmCompile> {
+                kotlinOptions {
+                    useIR = true
+                    jvmTarget = KScienceVersions.JVM_TARGET.toString()
+                }
+            }
+
+            extensions.findByType<JavaPluginExtension>()?.apply {
+                targetCompatibility = KScienceVersions.JVM_TARGET
+            }
+
+            tasks.apply {
+                withType<Test> {
+                    useJUnitPlatform()
+                }
+            }
         }
 
         pluginManager.withPlugin("org.jetbrains.kotlin.js") {
@@ -51,7 +62,11 @@ open class KScienceCommonPlugin : Plugin<Project> {
                 explicitApiWarning()
 
                 js(IR) {
-                    browser()
+                    browser{
+                        commonWebpackConfig {
+                            cssSupport.enabled = true
+                        }
+                    }
                 }
 
                 sourceSets["main"].apply {
@@ -86,7 +101,11 @@ open class KScienceCommonPlugin : Plugin<Project> {
                 }
 
                 js(IR) {
-                    browser()
+                    browser{
+                        commonWebpackConfig {
+                            cssSupport.enabled = true
+                        }
+                    }
                 }
 
                 sourceSets.invoke {
@@ -123,25 +142,23 @@ open class KScienceCommonPlugin : Plugin<Project> {
                 (tasks.findByName("jsProcessResources") as? Copy)?.apply {
                     fromJsDependencies("jsRuntimeClasspath")
                 }
-            }
-        }
 
-        afterEvaluate {
-            extensions.findByType<JavaPluginExtension>()?.apply {
-                targetCompatibility = KScienceVersions.JVM_TARGET
-            }
+                extensions.findByType<JavaPluginExtension>()?.apply {
+                    targetCompatibility = KScienceVersions.JVM_TARGET
+                }
 
-            tasks.apply {
-                withType<KotlinJvmCompile> {
-                    kotlinOptions {
-                        useIR = true
-                        jvmTarget = KScienceVersions.JVM_TARGET.toString()
+                tasks.apply {
+                    withType<Test> {
+                        useJUnitPlatform()
                     }
                 }
-                withType<Test> {
-                    useJUnitPlatform()
-                }
             }
         }
+
+        // apply dokka for all projects
+        if (!plugins.hasPlugin("org.jetbrains.dokka")) {
+            plugins.apply("org.jetbrains.dokka")
+        }
+
     }
 }