+++ src/main/kotlin/ru/mipt/npm/gradle/KScienceNativePlugin.kt
@@ -1,12 +1,20 @@
 package ru.mipt.npm.gradle
 
+import org.gradle.api.Action
+import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.findPlugin
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
+
+private fun KotlinMultiplatformExtension.sourceSets(configure: Action<NamedDomainObjectContainer<KotlinSourceSet>>): Unit =
+    (this as org.gradle.api.plugins.ExtensionAware).extensions.configure("sourceSets", configure)
 
 class KScienceNativePlugin : Plugin<Project> {
-    override fun apply(target: Project) = target.run {
+    override fun apply(project: Project) = project.run {
         //Apply multiplatform plugin is not applied, apply it
         if (plugins.findPlugin("org.jetbrains.kotlin.multiplatform") == null) {
             logger.info("Kotlin multiplatform plugin is not resolved. Adding it automatically")
@@ -18,46 +26,42 @@ class KScienceNativePlugin : Plugin<Project> {
         }
 
         configure<KotlinMultiplatformExtension> {
-            //deploy mode
-            linuxX64()
-            mingwX64()
-            macosX64()
+            val hostOs = System.getProperty("os.name")
+
+            val isLinux = hostOs == "Linux"
+            val isMinGw = hostOs.startsWith("Windows")
+            val isMacOs = hostOs == "Mac OS X"
+
+            if (isLinux || isMinGw) {
+                linuxX64()
+            }
+            if (isMinGw) {
+                mingwX64()
+            }
+            if (isMacOs) {
+                macosX64()
+            }
 
             sourceSets {
-                val commonMain by getting
-                val commonTest by getting
+                val commonMain = findByName("commonMain")!!
+                val commonTest = findByName("commonTest")!!
 
-                val nativeMain by creating {
+                val nativeMain = create("nativeMain").apply {
                     dependsOn(commonMain)
                 }
 
-                val nativeTest by creating {
+                val nativeTest = create("nativeTest").apply {
                     dependsOn(commonTest)
                 }
 
-                val linuxX64Main by getting {
-                    dependsOn(nativeMain)
-                }
+                findByName("linuxX64Main")?.dependsOn(nativeMain)
+                findByName("linuxX64Test")?.dependsOn(nativeTest)
 
-                val mingwX64Main by getting {
-                    dependsOn(nativeMain)
-                }
+                findByName("mingwX64Main")?.dependsOn(nativeMain)
+                findByName("mingwX64Test")?.dependsOn(nativeTest)
 
-                val macosX64Main by getting {
-                    dependsOn(nativeMain)
-                }
-
-                val linuxX64Test by getting {
-                    dependsOn(nativeTest)
-                }
-
-                val mingwX64Test by getting {
-                    dependsOn(nativeTest)
-                }
-
-                val macosX64Test by getting {
-                    dependsOn(nativeTest)
-                }
+                findByName("macosX64Main")?.dependsOn(nativeMain)
+                findByName("macosX64Test")?.dependsOn(nativeTest)
             }
         }
     }